refactor: centralize skill helper utilities
